+++ config/routes.rb
@@ -1,5 +1,5 @@
 BookingSync::Engine.routes.draw do
-  get '/auth/:provider/callback', to: 'sessions#create'
+  get '/auth/bookingsync/callback', to: 'sessions#create'
   get '/auth/failure', to: 'sessions#failure'
   get '/signout' => 'sessions#destroy', as: :signout
 end